Initial open source commit
The code in this commit is based off of the private version of Husky
Trails not yet released to the public. To open source the project, it
was rebased onto a new install of Laravel with UConn specific
information removed.

The project can now be used by other institutions, but some code cleanup
will be required before it's easy for them to do so.

Brandon